+++ ocr_translate/ocr_tsl/huggingface.py
